+++ Sources/WASIBase/Platform/Directory.swift
