+++ Sources/WASIBase/Platform/Entry.swift
