+++ Sources/WASIBase/Platform/File.swift
