+++ Sources/WASIBase/Platform/PlatformTypes.swift
