+++ Sources/WASIBase/Platform/SandboxPrimitives.swift
