+++ Sources/WasmKit/Execution/Types/Value.swift
@@ -54,26 +54,7 @@ extension WasmParser.ValueType {
 
 public typealias ReferenceType =  WasmParser.ReferenceType
 
-public enum Reference: Hashable {
-    /// A reference to a function.
-    case function(FunctionAddress?)
-    /// A reference to an external entity.
-    case extern(ExternAddress?)
-}
-
-/// Runtime representation of a value.
-public enum Value: Hashable {
-    /// Value of a 32-bit signed or unsigned integer.
-    case i32(UInt32)
-    /// Value of a 64-bit signed or unsigned integer.
-    case i64(UInt64)
-    /// Value of a 32-bit IEEE 754 floating-point number.
-    case f32(UInt32)
-    /// Value of a 64-bit IEEE 754 floating-point number.
-    case f64(UInt64)
-    /// Reference value.
-    case ref(Reference)
-
+extension Value {
     var type: WasmParser.ValueType {
         switch self {
         case .i32:
@@ -100,54 +81,6 @@ public enum Value: Hashable {
         default:
             fatalError("unknown raw integer type \(Swift.type(of: rawValue)) passed to `Value.init` ")
         }
-    }
-
-    /// Create a new value from a signed 32-bit integer.
-    public init(signed value: Int32) {
-        self = .i32(UInt32(bitPattern: value))
-    }
-
-    /// Create a new value from a signed 64-bit integer.
-    public init(signed value: Int64) {
-        self = .i64(UInt64(bitPattern: value))
-    }
-
-    /// Create a new value from a 32-bit floating-point number.
-    public static func fromFloat32(_ value: Float32) -> Value {
-        return .f32(value.bitPattern)
-    }
-
-    /// Create a new value from a 64-bit floating-point number.
-    public static func fromFloat64(_ value: Float64) -> Value {
-        return .f64(value.bitPattern)
-    }
-
-    /// Returns the value as a 32-bit signed integer.
-    /// - Precondition: The value is of type `i32`.
-    public var i32: UInt32 {
-        guard case let .i32(result) = self else { fatalError() }
-        return result
-    }
-
-    /// Returns the value as a 64-bit signed integer.
-    /// - Precondition: The value is of type `i64`.
-    public var i64: UInt64 {
-        guard case let .i64(result) = self else { fatalError() }
-        return result
-    }
-
-    /// Returns the value as a 32-bit floating-point number.
-    /// - Precondition: The value is of type `f32`.
-    public var f32: UInt32 {
-        guard case let .f32(result) = self else { fatalError() }
-        return result
-    }
-
-    /// Returns the value as a 64-bit floating-point number.
-    /// - Precondition: The value is of type `f64`.
-    public var f64: UInt64 {
-        guard case let .f64(result) = self else { fatalError() }
-        return result
     }
 
     func asAddressOffset(_ isMemory64: Bool) -> UInt64 {
